requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
argon2-cffi==23.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.6
